Plural pronouns for singular annoys me.
I really believe in gender fairness and fluidness in prounouning, and I
get why "they" is always better to use, and that the "one's request"
construction is too tortured.  So, I tend to just write everything in
plurals if possible which keeps 'they' but makes the number
grammatically agree.

Yes, I know that the ADT made "they/them used as singular" as 2015 word
of the year. (see
https://en.wikipedia.org/wiki/Word_of_the_year#American_Dialect_Society

so maybe some nun slapped the table too hard with her ruler when made
some plural/singular mistake with a pronoun when I was in grade school,
but I have a compulsive need to change everything to plural when I see
"they" and am editing a document. ;)