Ensure user freedom for Javascript.
The fundamental point here probably goes without saying given who the
project leader is. ;)

The LibreJS thing may end up to be nice-to-have.  LibreJS has some
serious problems -- I've had difficulty getting websites to work with
the plugin because the LibreJS plugin makes overly simplistic
assumptions about how Javascript is often deployed on a website.

But, we should try to be compatible if it's possible.