They aren't possible, but maybe are necessary.
I agree that all the items listed in here aren't possible given the time
allotted for development, and I haven't tried to integrate any of them
into the main feature set to avoid feature creep.  However, some of them
are going to be necessary, and early users may need to realize that some
out-of-band workarounds will be necessary in the first version, because
the features listed here were, well, necessary. :)