Policy validation can wait until after the first release.
The policy validations that would help reduce back and forth most are
the ones that are hard to implement: checking that attachments
actually include the necessary information, checking that per diems
match limits, etc.  Building to be able to accommodate these is going
to take time, and we don't need to make all that investment for the
first release.